requirements.txt
asgiref==3.8.1
cffi==1.17.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
packaging==24.2
pycparser==2.22
setuptools==75.4.0
sqlparse==0.5.1
tzdata==2024.2
zope.event==5.0
zope.interface==7.1.1
rest-framework-simplejwt==0.0.2
djangorestframework-simplejwt==5.3.1